Refactor application layer to domain entities
