+++ app/Application/Reprogramaciones/ReprogramacionService.php
@@ -6,8 +6,8 @@ use App\Application\Solicitudes\SolicitudService;
 use App\Domain\Reprogramacion\Repositories\ReprogramacionRepository;
 use App\Enums\EstadoAsistencia;
 use App\Mail\RescheduleMail;
-use App\Models\ModuloDocente\Reprogramacion;
-use App\Models\ModuloEstudiante\Solicitud;
+use App\Domain\Reprogramacion\Entities\Reprogramacion;
+use App\Domain\Solicitud\Entities\Solicitud;
 use Illuminate\Support\Facades\Mail;
 
 class ReprogramacionService
