+++ app/Domain/Catalogo/Entities/Asignatura.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace App\Models\ModuloSecretaria;
+namespace App\Domain\Catalogo\Entities;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
-// Models
-use App\Models\ModuloSecretaria\Facultad;
-use App\Models\ModuloEstudiante\Solicitud;
+use App\Domain\Catalogo\Entities\Facultad;
+use App\Domain\Solicitud\Entities\Solicitud;
 
 class Asignatura extends Model
 {
