+++ app/Domain/Seguridad/Entities/Role.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\ModuloSeguridad;
+namespace App\Domain\Seguridad\Entities;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
