+++ app/Infraestructure/Persistence/Eloquent/Repositories/EloquentApelacionRepository.php
@@ -4,7 +4,7 @@ namespace App\Infraestructure\Persistence\Eloquent\Repositories;
 
 use App\Domain\Apelaciones\Repositories\ApelacionRepository;
 use App\Enums\EstadoApelacion;
-use App\Models\ModuloEstudiante\Apelacion;
+use App\Domain\Apelaciones\Entities\Apelacion;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
