+++ app/Infraestructure/Persistence/Eloquent/Repositories/EloquentFacultadRepository.php
@@ -3,7 +3,7 @@
 namespace App\Infraestructure\Persistence\Eloquent\Repositories;
 
 use App\Domain\Catalogo\Repositories\FacultadRepository;
-use App\Models\ModuloSecretaria\Facultad;
+use App\Domain\Catalogo\Entities\Facultad;
 use Illuminate\Support\Collection;
 
 class EloquentFacultadRepository implements FacultadRepository
