+++ app/Infraestructure/Persistence/Eloquent/Repositories/EloquentReprogramacionRepository.php
@@ -3,7 +3,7 @@
 namespace App\Infraestructure\Persistence\Eloquent\Repositories;
 
 use App\Domain\Reprogramacion\Repositories\ReprogramacionRepository;
-use App\Models\ModuloDocente\Reprogramacion;
+use App\Domain\Reprogramacion\Entities\Reprogramacion;
 
 class EloquentReprogramacionRepository implements ReprogramacionRepository
 {
