+++ app/Infraestructure/Persistence/Eloquent/Repositories/EloquentSolicitudRepository.php
@@ -5,7 +5,7 @@ namespace App\Infraestructure\Persistence\Eloquent\Repositories;
 use App\Domain\Solicitud\Repositories\SolicitudRepository;
 use App\Enums\EstadoApelacion;
 use App\Enums\EstadoSolicitud;
-use App\Models\ModuloEstudiante\Solicitud;
+use App\Domain\Solicitud\Entities\Solicitud;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
