+++ app/Presentation/Http/Controllers/Auth/TeacherEmailVerificationOnLoginTest.php
@@ -2,8 +2,8 @@
 
 namespace Tests\Feature\Auth;
 
-use App\Models\User;
-use App\Models\ModuloSeguridad\Role;
+use App\Domain\Usuarios\Entities\User;
+use App\Domain\Seguridad\Entities\Role;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Notification;
 use Illuminate\Auth\Notifications\VerifyEmail;
