+++ app/Presentation/Http/Controllers/Secretaria/FacultadController.php
@@ -7,7 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
 
 // Models
-use App\Models\ModuloSecretaria\Facultad;
+use App\Domain\Catalogo\Entities\Facultad;
 
 class FacultadController extends Controller
 {
